CICD_AA/backend/requirements.txt
# Runtime dependencies for the CI/CD Healing Agent backend
# Pinned versions compatible with Python 3.11–3.14

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gitpython>=3.1.40
docker>=7.0.0
pytest>=8.0.0
pydantic>=2.5.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0

# LLM – NVIDIA API (OpenAI-compatible) + optional Ollama fallback
openai>=1.30.0

# Note: crewai has Python <3.12 restriction; we use concurrent.futures natively
# Uncomment if you are on Python 3.11:
# crewai>=0.28.0
# langchain>=0.2.0
# langchain-openai>=0.1.0
